Keep extraInfo and tags in object format for easy editing and allow specific keys to be removed
Previously, setting the same key multiple times would cause multiple values for that key in the array of pairs. It worked, but eventually this could result in a lot of networking overhead if tags or extra are frequently set. It was not possible to truly remove a key though setting its value to null or undefined may have had a similar effect. Updated .jshintrc to allow functions to be used before they are defined; allows utility functions to remain at the bottom of the script.